+++ rest-api/src/main/java/com/agenda/restapi/Contact/AccountController.java
@@ -18,13 +18,13 @@ public interface AccountController {
     AccountDto addAccount(@RequestBody AccountDto accountDto);
 
     @PutMapping("/edit/{id}")
-    AccountDto editAccount(@RequestBody AccountDto accountDto, @PathVariable Long id);
+    AccountDto editAccount(@RequestBody AccountDto accountDto, @PathVariable Integer id);
 
     @GetMapping("/find/{id}")
-    AccountDto findAccount(@PathVariable Long id);
+    AccountDto findAccount(@PathVariable Integer id);
 
     @DeleteMapping("/delete/{id}")
-    AccountDto deleteAccount(@PathVariable Long id);
+    AccountDto deleteAccount(@PathVariable Integer id);
 }
