+++ service-api-impl/src/main/java/com/agenda/serviceapiimpl/ContactServiceImpl.java
@@ -35,7 +35,7 @@ public class ContactServiceImpl implements ContactService {
     }
 
     @Override
-    public ContactDto getContactDTO(Long id) {
+    public ContactDto getContactDTO(Integer id) {
         if(contactRepository.findById(id).isPresent()){
             return contactRepository.findById(id).get().toContactDto();
         }
@@ -43,7 +43,7 @@ public class ContactServiceImpl implements ContactService {
     }
 
     @Override
-    public ContactDto editContactDTO(ContactDto contactDto, Long id) {
+    public ContactDto editContactDTO(ContactDto contactDto, Integer id) {
         Optional<Contact> dbContact = contactRepository.findById(id);
 
         if (dbContact.isPresent()) {
